+++ .gitignore
@@ -1,0 +1,4 @@
+# test ckpt and dataset
+public_data
+# hydra outputs 
+outputs
