+++ .vscode/launch.json
@@ -1,0 +1,16 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "test_LUNA",
+            "type": "python",
+            "request": "launch",
+            "program": "/workspace/SeqX2Y_PyTorch/test/test_LUNA.py",
+            "console": "integratedTerminal",
+            "justMyCode": true
+        }
+    ]
+}
