+++ project/models/__init__.py
@@ -1,0 +1,11 @@
+import sys
+
+try:
+    from ConvLSTMCell3d import * 
+    from layers import * 
+    from unet_utils import * 
+except:
+    sys.path.append('/workspace/SeqX2Y_PyTorch/project/models')
+    from ConvLSTMCell3d import * 
+    from layers import * 
+    from unet_utils import *
