+++ project/models/unet_model.py
@@ -1,13 +1,13 @@
+import os
 from PIL import Image
 import numpy as np
-import torch
 import matplotlib.pyplot as plt
+import torch
 from torch import nn
 import torch.nn.functional as F
-import os
 from glob import glob
 from layers import SpatialTransformer
-from models.unet_utils import *
+from unet_utils import *
