+++ test/__init__.py
@@ -1,0 +1,3 @@
+import sys
+sys.path.append('/workspace/SeqX2Y_PyTorch/project')
+sys.path.append('/workspace/SeqX2Y_PyTorch/')
